Termincacion de Crud con LINQ

+++ ML/Estado.cs
@@ -8,7 +8,7 @@ namespace ML
 {
     public class Estado
     {
-        public int IdEstado { get; set; }
+        public byte IdEstado { get; set; }
         public string Nombre { get; set; }
         public List<object> Estados { get; set; }
